Implement ten new methods in the Iter class.
- [x] Improves map_intersperse
- [x] Implements map_join
- [x] Implements map_reduce
- [x] Implements max (max_by in Elixir)
- [x] Implements member
- [x] Implements min (min_by in Elixir)
- [x] Implements min_max
- [x] Implements product
- [x] Implements random
- [x] Implements reduce
- [x] Implements reduce_while

Remarks: There's doesn't seem to be a sorter parameter for the max in min
built-in functions in Python, so for the moment I have resorted to rename
max_by to max to stay close to Python's built-ins in this instance. The
random function essentially just invokes random.choice from the standard
library which is sufficiently random for the time being. As for the product
function, it's equivalent for x_1 * x_2 * x_k * ... * x_b for k in range [a, b],
which is different from itertools.product which yields the cartesian product
of two sets. The member function uses the <in> operator; for dictionary
specifically this means it checks keys for memberships.